Rebuild http/2 responsiveness server app
Motivation:

Build a basic implementation of the responsiveness server. It does not
need HTTP/1 or TLS to begin with, just HTTP/2. It should also support a
shutdown handler for automatic testing.

Modifications:

Rebuild the initial app to support just HTTP/2. Upgrade the
responsiveness server to the new version, which supports a fallback
handler to implement the shutdown endpoint. Integrate the
ServerQuiescingHelper and connection management to properly shutdown the
server.

Result:

The new server only supports HTTP/2 and can be shut down gracefully.